Remove standalone feature from runtime.
There is truly only one runtime now.
